Changed genereId type to int

+++ Vidly/Models/Customer.cs
@@ -20,7 +20,7 @@ namespace Vidly.Models
         public MembershipType MembershipType { get; set; }
 
         [Display(Name = "Membership Type")]
-        public byte MembershipTypeId { get; set; }
+        public int MembershipTypeId { get; set; }
 
         [Display(Name ="Date of birth")]
         public DateTime? Birthdate { get; set; }
